# Speed up HNSW index build