@hlinnaka 2023-10-04
> Nice! For sanity checking, did you do any recall testing as well?

No, but there is that assertion that it makes the same decisions as on 'master', so recall should be identical. Makes sense as a sanity check though; any tips on how to run such a test?