@ankane 2023-10-04
Hi @hlinnaka, thanks for the PR! For the first algorithm, I think this could be done by keeping track of the first pruned candidate for each layer, which would require less memory and minimize complexity. Will try to experiment with it.