@ankane 2023-10-04
Need to do some more testing, but pushed roughly what I'm thinking to the [hnsw-build-speedup branch](https://github.com/pgvector/pgvector/compare/hnsw-build-speedup).