@ankane 2023-10-05
My bad, it shouldn't take up any more memory on 64-bit machines due to alignment. Just pushed an updated version that should get all the performance benefits of the full patchset. Let me know what you think.