@hlinnaka 2023-10-05
> Hi @hlinnaka, thanks for the PR! For the first algorithm, I think this could be done by keeping track of the first pruned candidate for each layer, which would require less memory and minimize complexity. Will try to experiment with it.

Yeah, that would work too. I'm not sure if the neighbors array is currently sorted such that the non-pruned elements come before the pruned candidates, but we could probably arrange that.

> My bad, it shouldn't take up any more memory on 64-bit machines due to alignment.

Good point. The opportunity cost is that we could pack the arrays more tightly to save more memory, though.

> Just pushed an updated version that should get all the performance benefits of the full patchset. Let me know what you think.

It's pretty close to the 2nd patch in this PR, but doesn't get the benefits of the 3rd patch. Compared to the 2nd patch here, it missed one trick: If the new element is not "closer", then the rest of the closer elements don't need to be recalculated:

```
diff --git a/src/hnswutils.c b/src/hnswutils.c
index 5d0743c..801aa90 100644
--- a/src/hnswutils.c
+++ b/src/hnswutils.c
@@ -811,7 +811,7 @@ SelectNeighbors(List *c, int m, int lc, FmgrInfo *procinfo, Oid collation, HnswE
                else if (e == newCandidate)
                {
                        e->closer = CheckElementCloser(e, r, lc, procinfo, collation);
-                       mustCalculate = true;
+                       mustCalculate = e->closer;
                }
 
                if (e->closer)
```

With that change, it has the same effect as the 2nd patch in this PR. +1 for that.

The 3rd patch is smarter about recalculating the elements after the new element has been added to the closer set. The key insight is that after you add an element to the closer set, you don't need to fully re-evaluate the elements that come after that. Any element that was not in the closer set before still won't be in the closer set. And any element that was previously in the closer set, it's sufficient to check if they are closer to the newly-added vector, because adding a new vector doesn't change the distance to any of the other vectors in the set. Only after you _remove_ a vector from the closer set, you need to fully recalculate.