@ankane 2023-10-05
Thanks @hlinnaka, really appreciate the great explanation (missed both those). Just pushed a new commit that should address them (trying to keep the changeset minimal). Let me know if it looks good now (the performance win is incredible).