@hlinnaka 2023-10-05
LGTM, thank you!

To test this, I used the following little patch to check that SelectNeighbors makes the same decisions incrementally as when it recalculates from scratch:

```
diff --git a/src/hnswutils.c b/src/hnswutils.c
index e270541..62537eb 100644
--- a/src/hnswutils.c
+++ b/src/hnswutils.c
@@ -961,6 +961,7 @@ HnswUpdateConnection(HnswElement element, HnswCandidate * hc, int m, int lc, int
 		if (pruned == NULL)
 		{
 			List	   *c = NIL;
+			HnswCandidate *prunedVerify = NULL;
 
 			/* Add and sort candidates */
 			for (int i = 0; i < currentNeighbors->length; i++)
@@ -968,8 +969,35 @@ HnswUpdateConnection(HnswElement element, HnswCandidate * hc, int m, int lc, int
 			c = lappend(c, &hc2);
 			list_sort(c, CompareCandidateDistances);
 
+			if (hc->element->neighbors[lc].closerSet)
+			{
+				/*
+				 * Make a copy of the input list 'c' for verification. (SelectNeighbors
+				 * modifies it in place)
+				 */
+				ListCell *cell;
+				List *c_copy = NIL;
+
+				foreach (cell, c)
+				{
+					HnswCandidate *orig = (HnswCandidate *) lfirst(cell);
+					HnswCandidate *copy = palloc(sizeof(HnswCandidate));
+
+					memcpy(copy, orig, sizeof(HnswCandidate));
+					c_copy = lappend(c_copy, copy);
+				}
+
+				/* temporarily force SelectNeighbors to recalculate the set */
+				hc->element->neighbors[lc].closerSet = false;
+				SelectNeighbors(c_copy, m, lc, procinfo, collation, hc->element, NULL, &prunedVerify);
+				hc->element->neighbors[lc].closerSet = true;
+			}
+
 			SelectNeighbors(c, m, lc, procinfo, collation, hc->element, &hc2, &pruned);
 
+			if (prunedVerify && prunedVerify->element != pruned->element)
+				elog(PANIC, "Nope");
+
 			/* Should not happen */
 			if (pruned == NULL)
 				return;
```
The regression tests passed with that, as well as some manual testing with the dataset I used earlier.