@ankane 2023-10-05
Awesome, that's a good check. Thanks for verifying!

It looks like the performance benefit is even greater with more dimensions (which makes sense). I'm seeing 1536 dimensions take 1/3 of the time to build locally.

Will release with 0.5.1 once #260 is resolved. This is a great win, so big thanks!!