@ankane 2023-10-06
fwiw, seeing failures in the TAP tests with the verification code (with both this and #295) that aren't present when `mustCalculate` is hard-coded to `true`. Still looking into it.