@ankane 2023-10-06
Fixed in cae162ffc67aefa139176d2a6f0f0f53cf1f80ab, as the logic relies on the candidate list having the same order between calls.