@hlinnaka 2023-10-06
> Fixed in [cae162f](https://github.com/pgvector/pgvector/commit/cae162ffc67aefa139176d2a6f0f0f53cf1f80ab), as the logic relies on the candidate list having the same order between calls.

Hmm, is that needed just to make it deterministic so that you get the same results as with the verification code pass, or does it actually affect the correctness?