@ankane 2023-10-06
It affects correctness, as the cached value of `closer` is no longer valid if two elements with the same distance are swapped.