@ankane 2023-10-06
One way to confirm is by changing the check to:

```c
if (prunedVerify && prunedVerify->element != pruned->element && pruned->distance != prunedVerify->distance)
    elog(PANIC, "Nope %f = %f", pruned->distance, prunedVerify->distance);
```

And running the `011_hnsw_vacuum.pl` test in a loop. On some runs, it'll end up selecting an element with a different distance to prune.